moved QuantumHeadHamiltonian classes to separate script
